# [8.x] Optimize PHP performance by using fully-qualified function calls