--- a/src/Illuminate/Auth/AuthServiceProvider.php
+++ b/src/Illuminate/Auth/AuthServiceProvider.php
@@ -51,7 +51,7 @@ class AuthServiceProvider extends ServiceProvider
     protected function registerUserResolver()
     {
         $this->app->bind(AuthenticatableContract::class, function ($app) {
-            return call_user_func($app['auth']->userResolver());
+            return \call_user_func($app['auth']->userResolver());
         });
     }
 
@@ -64,7 +64,7 @@ class AuthServiceProvider extends ServiceProvider
     {
         $this->app->singleton(GateContract::class, function ($app) {
             return new Gate($app, function () use ($app) {
-                return call_user_func($app['auth']->userResolver());
+                return \call_user_func($app['auth']->userResolver());
             });
         });
     }
@@ -94,7 +94,7 @@ class AuthServiceProvider extends ServiceProvider
     {
         $this->app->rebinding('request', function ($app, $request) {
             $request->setUserResolver(function ($guard = null) use ($app) {
-                return call_user_func($app['auth']->userResolver(), $guard);
+                return \call_user_func($app['auth']->userResolver(), $guard);
             });
         });
     }