--- a/src/Illuminate/Auth/Middleware/Authorize.php
+++ b/src/Illuminate/Auth/Middleware/Authorize.php
@@ -54,7 +54,7 @@ class Authorize
      */
     protected function getGateArguments($request, $models)
     {
-        if (is_null($models)) {
+        if (\is_null($models)) {
             return [];
         }