--- a/src/Illuminate/Auth/Passwords/PasswordBroker.php
+++ b/src/Illuminate/Auth/Passwords/PasswordBroker.php
@@ -52,7 +52,7 @@ class PasswordBroker implements PasswordBrokerContract
         // "flash" data in the session to indicate to the developers the errors.
         $user = $this->getUser($credentials);
 
-        if (is_null($user)) {
+        if (\is_null($user)) {
             return static::INVALID_USER;
         }
 
@@ -112,7 +112,7 @@ class PasswordBroker implements PasswordBrokerContract
      */
     protected function validateReset(array $credentials)
     {
-        if (is_null($user = $this->getUser($credentials))) {
+        if (\is_null($user = $this->getUser($credentials))) {
             return static::INVALID_USER;
         }