--- a/src/Illuminate/Auth/RequestGuard.php
+++ b/src/Illuminate/Auth/RequestGuard.php
@@ -50,11 +50,11 @@ class RequestGuard implements Guard
         // If we've already retrieved the user for the current request we can just
         // return it back immediately. We do not want to fetch the user data on
         // every call to this method because that would be tremendously slow.
-        if (! is_null($this->user)) {
+        if (! \is_null($this->user)) {
             return $this->user;
         }
 
-        return $this->user = call_user_func(
+        return $this->user = \call_user_func(
             $this->callback, $this->request, $this->getProvider()
         );
     }
@@ -67,7 +67,7 @@ class RequestGuard implements Guard
      */
     public function validate(array $credentials = [])
     {
-        return ! is_null((new static(
+        return ! \is_null((new static(
             $this->callback, $credentials['request'], $this->getProvider()
         ))->user());
     }