--- a/src/Illuminate/Broadcasting/Broadcasters/Broadcaster.php
+++ b/src/Illuminate/Broadcasting/Broadcasters/Broadcaster.php
@@ -111,9 +111,9 @@ abstract class Broadcaster implements BroadcasterContract
      */
     protected function extractParameters($callback)
     {
-        if (is_callable($callback)) {
+        if (\is_callable($callback)) {
             return (new ReflectionFunction($callback))->getParameters();
-        } elseif (is_string($callback)) {
+        } elseif (\is_string($callback)) {
             return $this->extractParametersFromClass($callback);
         }
 
@@ -182,7 +182,7 @@ abstract class Broadcaster implements BroadcasterContract
         $binder = $this->binder();
 
         if ($binder && $binder->getBindingCallback($key)) {
-            return call_user_func($binder->getBindingCallback($key), $value);
+            return \call_user_func($binder->getBindingCallback($key), $value);
         }
 
         return $value;
@@ -207,7 +207,7 @@ abstract class Broadcaster implements BroadcasterContract
 
             $className = Reflector::getParameterClassName($parameter);
 
-            if (is_null($model = (new $className)->resolveRouteBinding($value))) {
+            if (\is_null($model = (new $className)->resolveRouteBinding($value))) {
                 throw new AccessDeniedHttpException;
             }
 
@@ -266,7 +266,7 @@ abstract class Broadcaster implements BroadcasterContract
      */
     protected function normalizeChannelHandlerToCallable($callback)
     {
-        return is_callable($callback) ? $callback : function (...$args) use ($callback) {
+        return \is_callable($callback) ? $callback : function (...$args) use ($callback) {
             return Container::getInstance()
                 ->make($callback)
                 ->join(...$args);
@@ -286,7 +286,7 @@ abstract class Broadcaster implements BroadcasterContract
 
         $guards = $options['guards'] ?? null;
 
-        if (is_null($guards)) {
+        if (\is_null($guards)) {
             return $request->user();
         }