--- a/src/Illuminate/Broadcasting/Broadcasters/RedisBroadcaster.php
+++ b/src/Illuminate/Broadcasting/Broadcasters/RedisBroadcaster.php
@@ -80,7 +80,7 @@ class RedisBroadcaster extends Broadcaster
      */
     public function validAuthenticationResponse($request, $result)
     {
-        if (is_bool($result)) {
+        if (\is_bool($result)) {
             return json_encode($result);
         }