--- a/src/Illuminate/Bus/Batch.php
+++ b/src/Illuminate/Bus/Batch.php
@@ -166,8 +166,8 @@ class Batch implements Arrayable, JsonSerializable
         $jobs = Collection::wrap($jobs)->map(function ($job) use (&$count) {
             $job = $job instanceof Closure ? CallQueuedClosure::create($job) : $job;
 
-            if (is_array($job)) {
-                $count += count($job);
+            if (\is_array($job)) {
+                $count += \count($job);
 
                 return with($this->prepareBatchedChain($job), function ($chain) {
                     return $chain->first()
@@ -281,7 +281,7 @@ class Batch implements Arrayable, JsonSerializable
      */
     public function finished()
     {
-        return ! is_null($this->finishedAt);
+        return ! \is_null($this->finishedAt);
     }
 
     /**
@@ -404,7 +404,7 @@ class Batch implements Arrayable, JsonSerializable
      */
     public function cancelled()
     {
-        return ! is_null($this->cancelledAt);
+        return ! \is_null($this->cancelledAt);
     }
 
     /**
@@ -425,11 +425,11 @@ class Batch implements Arrayable, JsonSerializable
      * @param  \Throwable|null  $e
      * @return void
      */
-    protected function invokeHandlerCallback($handler, Batch $batch, Throwable $e = null)
+    protected function invokeHandlerCallback($handler, self $batch, Throwable $e = null)
     {
         return $handler instanceof SerializableClosure
                     ? $handler->__invoke($batch, $e)
-                    : call_user_func($handler, $batch, $e);
+                    : \call_user_func($handler, $batch, $e);
     }
 
     /**