--- a/src/Illuminate/Cache/ApcWrapper.php
+++ b/src/Illuminate/Cache/ApcWrapper.php
@@ -18,7 +18,7 @@ class ApcWrapper
      */
     public function __construct()
     {
-        $this->apcu = function_exists('apcu_fetch');
+        $this->apcu = \function_exists('apcu_fetch');
     }
 
     /**