… .styleci.yml
… src/Illuminate/Bus/Dispatcher.php
… src/Illuminate/Cache/ArrayStore.php
… src/Illuminate/Cache/CacheManager.php
… src/Illuminate/Cache/DatabaseStore.php
… src/Illuminate/Cache/FileStore.php
… src/Illuminate/Cache/Lock.php
… src/Illuminate/Cache/MemcachedConnector.php
… src/Illuminate/Cache/PhpRedisLock.php
… src/Illuminate/Cache/RedisStore.php
… src/Illuminate/Cache/RedisTaggedCache.php
… src/Illuminate/Cache/Repository.php
… src/Illuminate/Cache/RetrievesMultipleKeys.php
… src/Illuminate/Cache/TaggableStore.php
… src/Illuminate/Collections/Arr.php
… src/Illuminate/Collections/Collection.php
… src/Illuminate/Collections/HigherOrderCollectionProxy.php
… src/Illuminate/Collections/LazyCollection.php
… src/Illuminate/Collections/Traits/EnumeratesValues.php
… src/Illuminate/Config/Repository.php
… src/Illuminate/Console/Application.php
… src/Illuminate/Console/Concerns/InteractsWithIO.php
… src/Illuminate/Console/ConfirmableTrait.php
… src/Illuminate/Console/GeneratorCommand.php
… src/Illuminate/Console/Parser.php
… src/Illuminate/Console/Scheduling/CallbackEvent.php
… src/Illuminate/Console/Scheduling/Event.php
… src/Illuminate/Console/Scheduling/ManagesFrequencies.php
… src/Illuminate/Console/Scheduling/Schedule.php
… src/Illuminate/Container/BoundMethod.php