This PR enables StyleCI's [`native_function_invocation_symfony`](https://docs.styleci.io/fixers#native_function_invocation_symfony) fixer & introduces changes to the way native PHP functions are referenced.

By directly referencing these functions in their (global) namespace (`\is_null`) instead of calling them within the current scope (`is_null`), PHP can skip the step of resolving this function within the current namespace. More importantly, it also allows PHP's OPCache to cache these calls, which leads to significant performance improvements.

We've benchmarked these changes on a [Forge](https://forge.laravel.com) provisioned server on DigitalOcean, running a similar setup [as described here](https://medium.com/@taylorotwell/benchmarking-laravel-symfony-zend-2c01c2b270f8). Just to recap, this means our configuration is looking like this:

- PHP 7.4, with OPCache enabled.
- `web` (session) middleware (`app/Http/Kernel.php`) disabled.
- Optimized composer autoloader (`composer dumpautoload -o`)
- Cached configuration (`php artisan config:cache`)

Furthermore, to benchmark, we're running `ab -t 10 -c 10 http://server.address/` against the homepage (`/`) of an otherwise fresh [`laravel/laravel`](https://github.com/laravel/laravel) installation.

To make sure that we weren't catching the server in a bad moment, we ran the benchmark five times each at least a minute apart, and gathered the **mean** requests per second of every run:

| # | Before | After |
| -- | -- | -- |
| 1 | 801,77 | 937,17 |
| 2 | 717,24 | 681,92 |
| 3 | 587,44 | 689,93 |
| 4 | 554,62 | 674,22 |
| 5 | 469,30 | 947,28 |
| 🔥  | **626,074** | **786,104** |

That's a **25.56%** performance improvement on a fresh Laravel v8.44.0
When introducing a database to the mix, the gains drop dramatically to only 2,63% (see https://github.com/laravel/framework/pull/37517#issuecomment-853062134)

(Inspired by https://github.com/scssphp/scssphp/issues/96)