@eusonlito 2021-05-28
Only as comment, on my deploys I have a `php-cs-fixer` command to fix all `vendor` folder with `native_constant_invocation` and `native_function_invocation`:

### .php-cs-fixer.php

```php
<?php declare(strict_types=1);

use PhpCsFixer\Finder;
use PhpCsFixer\Config;

return (new Config())
    ->setRiskyAllowed(true)
    ->setRules(['native_constant_invocation' => true, 'native_function_invocation' => true])
    ->setFinder(Finder::create()->exclude(['phpunit', 'php-parallel-lint', 'tests'])->in('vendor'));
```

### Execution inside deploy script:

```
$> php ./vendor/bin/php-cs-fixer fix --ansi --quiet
```

The performance difference is very noticeable.