@s-patompong 2021-05-28
> @eusonlito That is a very risky thing to do, and will break octan, for example.

Yeah, the doc also mentioned this:
```
This is a "risky" heuristic fixer, and could change code behavior!
```