@eusonlito 2021-05-28
> @eusonlito That is a very risky thing to do, and will break octan, for example.

@GrahamCampbell thanks for your advice. I haven't tried Octane yet, but in about 5 years using this method in about 50 or 60 Laravel projects, never notice any issue.

Anyway, thanks. I will check it.