@eusonlito 2021-05-28
> It specifically breaks the ability to override core functions. Uploading files over HTTP in octane will break if you run that command on the vendor directory.

If my memory doesn't fail, I think that Guzzle had same *problem* on version 6 and was solved on code at version 7.