@deleugpn 2021-05-28
Is it a boost enough to justify the ugly codebase though? 🤷‍♂️