@claudiodekker 2021-05-28
> But is it a boost enough to justify the ugly codebase though? 🤷‍♂️

Yeah, that's indeed the question.. Internally, I already mentioned that it'd be great if there was a way to have these be auto-imported when referenced (`use \is_null;`), as then it'd only occur a few time near the top of the file. In theory, that should also be just-as-quick.

> Can anyone explain why it cant be done in php core itself ? I mean can't they add performance by default ?

They probably can, this is just one remnants from an older PHP version. It's kinda silly (IMO) to re-define built-in langauge functions in your local namespace anyway ¯\_(ツ)_/¯