@timacdonald 2021-05-29
> Yeah, that's indeed the question.. Internally, I already mentioned that it'd be great if there was a way to have these be auto-imported when referenced (use \is_null;), as then it'd only occur a few time near the top of the file. In theory, that should also be just-as-quick.

Totally possible and something I do on my projects instead of messing with the app code itself.

See: https://github.com/timacdonald/log-fake/blob/master/src/LogFake.php#L11

Via the PHP-CS-Fixer rule: https://github.com/FriendsOfPHP/PHP-CS-Fixer/blob/3.0/doc/rules/import/global_namespace_import.rst

I'd really prefer to see that implementation personally and keep the mess at the top of the file rather than scattered throughout.