@taylorotwell 2021-05-29
Do we still see any measurable performance benefit when adding in for example 1 Eloquent DB call into the request lifecycle?