@Krisell 2021-05-29
> Do we still see any measurable performance benefit when adding in for example 1 Eloquent DB call into the request lifecycle?

Waiting for DB-response doesn't use CPU-time so even if the difference is much smaller, or even insignificant, that doesn't mean the cpu load decrease is also insignificant. But it would definitely be interesting to see the numbers.