@garygreen 2021-05-31
> [in reference to having this be built into core PHP] That said, IMO it's still kinda silly to even want the ability to override built-in language functions within your own namespaces, but there's probably a real use-case that I'm not aware of out there somewhere ¯_(ツ)_/¯

I totally agree it does seem like every PHP codebase is being slowed down for that 0.01% edge case who override built in php functions based on namespaces.

Rather than importing for each file as you suggested `use \is_null` etc instead just cater to those edge cases who want to override php builtins `declare(override_builtins = 1);` - suggest default value should be `0`. That way `\is_null` etc is not needed by default for any PHP class.

Would love to see an RFC.