@taylorotwell 2021-06-01
If someone could do some benchmarks on an application that makes 1 database query please let me know.