@eusonlito 2021-06-01
> If someone could do some benchmarks on an application that makes 1 database query please let me know.

I have created some "advanced" tests: https://gist.github.com/eusonlito/961e6e0a8dde3736fca5465be6fc2c64