@taylorotwell 2021-06-02
@eusonlito sorry - a lot of information there.

I am primarily interested in what is the total execution time of handling an HTTP request before and after this change when hitting a route with 1 database query.