@eusonlito 2021-06-02
@taylorotwell.

Here:

```
before - http://localhost/laravel/public/test/query/1

Requests per second:    882.14 [#/sec] (mean)
Time per request:       11.336 [ms] (mean)
Time per request:       1.134 [ms] (mean, across all concurrent requests)
```
```
after - http://localhost/laravel/public/test/query/1

Requests per second:    901.76 [#/sec] (mean)
Time per request:       11.089 [ms] (mean)
Time per request:       1.109 [ms] (mean, across all concurrent requests)