@garygreen 2021-06-02
That's only a 2% speed difference which puts it in the region of a micro optimization?

I am kind of surprised seeing @claudiodekker result of 25% speed difference. I'm highly sceptical this has skipped the attention of PHP maintainers, as such a speed increase would be an easy win for most PHP projects if this was configurable in some way.

Is it a possibility the original results are down to testing method, bad timing when running tests, other factors at play?