@claudiodekker 2021-06-02
No, the original results are accurate. I'm about to run another test with a single DB query against both. 
Will put both instances up so people can verify the results themselves.