@claudiodekker 2021-06-02
Updated results, using the DB branches I referred to in my previous comment:

| # | Before | After |
| -- | -- | -- |
| 1 | 353,49 | 364,29 |
| 2 | 363,89 | 360,46 |
| 3 | 356,39 | 374,08 |
| 4 | 353,76 | 358,54 |
| 5 | 357,18 | 374,35 |
| 🔥  | **356,942** | **366,344** |

As @garygreen & @eusonlito's results indeed already conclude, the moment a database is introduced, the gains drops all the way down to only **2,63%**.