@taylorotwell 2021-06-02
@eusonlito @claudiodekker 

Was this with or without opcache enabled?