@claudiodekker 2021-06-02
@taylorotwell With OPCache enabled, as are all the tests above. Same exact conditions as the original post.